+++ src/memory/ft_print_memory.c
@@ -6,13 +6,13 @@
 /*   By: abezanni <abezanni@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/31 22:33:54 by abezanni          #+#    #+#             */
-/*   Updated: 2018/10/21 17:39:43 by abezanni         ###   ########.fr       */
+/*   Updated: 2018/11/29 19:24:18 by abezanni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "basics.h"
 
-static void	init(int *i, int *j, char *str)
+static void	init(size_t *i, size_t *j, char *str)
 {
 	*i = 0;
 	*j = 40;
@@ -23,8 +23,8 @@ void		ft_print_memory(const void *addr, size_t size)
 {
 	unsigned char	*str;
 	char			tabmem[57];
-	int				i;
-	int				j;
+	size_t			i;
+	size_t			j;
 
 	init(&i, &j, tabmem);
 	str = (unsigned char *)addr;
