+++ src/num/ft_iterative_power.c
@@ -6,15 +6,15 @@
 /*   By: abezanni <abezanni@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/18 14:22:13 by abezanni          #+#    #+#             */
-/*   Updated: 2017/12/18 14:22:21 by abezanni         ###   ########.fr       */
+/*   Updated: 2018/11/29 19:20:36 by abezanni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <string.h>
 
-int	ft_iterative_power(int nb, int power)
+size_t	ft_iterative_power(int nb, int power)
 {
-	int back;
+	size_t back;
 
 	if (power < 0)
 		back = 0;
