+++ src/str/ft_split_charset.c
@@ -3,20 +3,21 @@
 /*                                                        :::      ::::::::   */
 /*   ft_split_charset.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ccoupez <ccoupez@student.42.fr>            +#+  +:+       +#+        */
+/*   By: abezanni <abezanni@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/04/18 15:15:24 by abezanni          #+#    #+#             */
-/*   Updated: 2018/06/06 14:07:45 by ccoupez          ###   ########.fr       */
+/*   Updated: 2018/11/29 19:14:32 by abezanni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "basics.h"
+#include <stdlib.h>
 
 char	**ft_split_charset(char *str, char *charset)
 {
 	char	**back;
-	int		size_words;
-	int		i;
+	size_t	size_words;
+	size_t	i;
 
 	if (!(str && charset) || !(back = malloc(sizeof(char*) *
 		((ft_nbr_words_charset(str, charset) + 1)))))
