+++ src/str/ft_stpcpy.c
@@ -6,15 +6,15 @@
 /*   By: abezanni <abezanni@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/06 17:16:24 by abezanni          #+#    #+#             */
-/*   Updated: 2018/10/04 18:37:29 by abezanni         ###   ########.fr       */
+/*   Updated: 2018/11/29 18:49:56 by abezanni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <string.h>
 
 char	*ft_stpcpy(char *dest, const char *src)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (src[i] != '\0')
