+++ src/str/ft_strstr.c
@@ -6,11 +6,11 @@
 /*   By: abezanni <abezanni@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/06 17:23:30 by abezanni          #+#    #+#             */
-/*   Updated: 2018/05/04 10:33:11 by abezanni         ###   ########.fr       */
+/*   Updated: 2018/11/29 19:01:04 by abezanni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <string.h>
 
 /*
 **	Cherche la chaine to_find dans la chaine str
@@ -18,8 +18,8 @@
 
 char	*ft_strstr(const char *str, const char *to_find)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = 0;
 	if (to_find[0] == '\0')
